refactor(wrapper): seperated logic for each feature of sherpa-onnx
